requirements.txt
flask 
onnxruntime-gpu
loguru
